# Let the UDP sink use a dispatcher thread