--- a/lib/statsd/instrument/udp_sink.rb
+++ b/lib/statsd/instrument/udp_sink.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 
+require 'objspace'
+require 'thread'
+
 # @note This class is part of the new Client implementation that is intended
 #   to become the new default in the next major release of this library.
 class StatsD::Instrument::UDPSink
@@ -8,13 +11,62 @@ class StatsD::Instrument::UDPSink
     new(host, Integer(port_as_string))
   end
 
+  class Dispatcher
+    def initialize(host, port, queue)
+      @host = host
+      @port = port
+      @queue = queue
+      @socket = nil
+      @thread = nil
+      ObjectSpace.define_finalizer(self, Proc.new { |o| o.stop })
+    end
+
+    def start
+      @thread = Thread.new { dispatcher_loop }
+      self
+    end
+
+    def stop
+      @queue.close
+      @thread.join(0.1)
+    rescue ThreadError
+    end
+
+    private
+    def dispatcher_loop
+      loop do
+        datagram = @queue.pop(false)
+        begin
+          socket.send(datagram, 0)
+        rescue ThreadError
+          socket.send(datagram, 0)
+        rescue SocketError, IOError, SystemCallError
+          # TODO: log?
+          invalidate_socket
+        end
+      end
+    end
+
+    def socket
+      if @socket.nil?
+        @socket = UDPSocket.new
+        @socket.connect(@host, @port)
+      end
+      @socket
+    end
+
+    def invalidate_socket
+      @socket = nil
+    end
+  end
+
   attr_reader :host, :port
 
   def initialize(host, port)
     @host = host
     @port = port
-    @mutex = Mutex.new
-    @socket = nil
+    @queue = Queue.new
+    @dispatcher = Dispatcher.new(host, port, @queue).start
   end
 
   def sample?(sample_rate)
@@ -22,41 +74,11 @@ class StatsD::Instrument::UDPSink
   end
 
   def <<(datagram)
-    with_socket { |socket| socket.send(datagram, 0) > 0 }
+    @queue << datagram
     self
-
-  rescue ThreadError
-    # In cases where a TERM or KILL signal has been sent, and we send stats as
-    # part of a signal handler, locks cannot be acquired, so we do our best
-    # to try and send the datagram without a lock.
-    socket.send(datagram, 0) > 0
-
-  rescue SocketError, IOError, SystemCallError
-    # TODO: log?
-    invalidate_socket
   end
 
   def addr
     "#{host}:#{port}"
-  end
-
-  private
-
-  def with_socket
-    @mutex.synchronize { yield(socket) }
-  end
-
-  def socket
-    if @socket.nil?
-      @socket = UDPSocket.new
-      @socket.connect(@host, @port)
-    end
-    @socket
-  end
-
-  def invalidate_socket
-    @mutex.synchronize do
-      @socket = nil
-    end
   end
 end