### Why?

Although UDP uses a connectionless communication model, the following user space code paths aren't free:

* [Mutex](https://github.com/ruby/ruby/blob/115fec062ccf7c6d72c8d5f64b7a5d84c9fb2dd8/thread_sync.c#L237-L317) held on the socket
* Sending datagrams still incurs a [system call](https://github.com/ruby/ruby/blob/1663d347c993debf7ed83e11e291e7a21e14ed03/ext/socket/udpsocket.c#L199-L220)

Using a `Queue` as contract with a simple dispatch Thread reduce datagram dispatch for the producer to an [array push](https://github.com/ruby/ruby/blob/115fec062ccf7c6d72c8d5f64b7a5d84c9fb2dd8/thread_sync.c#L784) and a user space [thread interrupt](https://github.com/ruby/ruby/blob/0256e4f0f5e10f0a15cbba2cd64e252dfa864e4a/vm_core.h#L1791 ) 

Master:

```
methodmissing:statsd-instrument lourens$ benchmark/send-metrics-to-local-udp-receiver
Warming up --------------------------------------
StatsD metrics to local UDP receiver (branch: master, sha: 07fb158)
                         1.213k i/100ms
Calculating -------------------------------------
StatsD metrics to local UDP receiver (branch: master, sha: 07fb158)
                         11.876k (± 3.9%) i/s -     59.437k in   5.012616s

To compare the performance of this revision against another revision (e.g. master),
check out a different branch and run this benchmark script again.
```

This PR:

```
methodmissing:statsd-instrument lourens$ benchmark/send-metrics-to-local-udp-receiver
Warming up --------------------------------------
StatsD metrics to local UDP receiver (branch: dispatch-thread, sha: 3137f7e)
                         4.199k i/100ms
Calculating -------------------------------------
StatsD metrics to local UDP receiver (branch: dispatch-thread, sha: 3137f7e)
                         44.251k (±25.9%) i/s -    201.552k in   5.042619s

Comparison:
StatsD metrics to local UDP receiver (branch: dispatch-thread, sha: 3137f7e):    44251.3 i/s
StatsD metrics to local UDP receiver (branch: master, sha: 07fb158):    11921.7 i/s - 3.71x  (± 0.00) slower
```

### Why not?

Threads 😄 The dispatcher is very thin and cleanly separates IO and socket init + teardown on error from the metrics producer thread. Cleanup is handled with a GC finalizer (which of course only fires on clean exit, but this is true for ObjectSpace in general too). See the throwaway section below

### Why this is throwaway

Queues do not work X process because https://github.com/ruby/ruby/blob/115fec062ccf7c6d72c8d5f64b7a5d84c9fb2dd8/thread_sync.c#L606-L619 and therefore test https://github.com/Shopify/statsd-instrument/blob/master/test/udp_sink_test.rb#L63-L82 would always fail because https://github.com/Shopify/statsd-instrument/blob/master/test/udp_sink_test.rb#L67 is enq'ed but `num_waiting` is reset to 0, with the dispatcher thread effectively never seeing the datagram as the queue has no waiters.