@casperisfine 2020-03-18
I love the idea. However I wonder if it would be worth having a single handler rather than one per udpsink.

But maybe that's not worth it as it's unlikelt to have multiple sink instances at the same time?