@csfrancis 2020-03-18
> The Queue instance would still involve a mutex / synchronization point, right? So the perf difference comes from making the syscall async.

No, there's no mutex. The difference is that for MRI, the GVL is preventing multiple threads from running at the same time. The `Queue#push` will trigger an MRI interrupt, which wakes up a thread that is blocking via `Queue#pop`: https://github.com/ruby/ruby/blob/115fec062ccf7c6d72c8d5f64b7a5d84c9fb2dd8/thread_sync.c#L785. This all happens in user-space without a syscall.