@dylanahsmith 2020-03-18
> No, there's no mutex. [...]. This all happens in user-space without a syscall.

If another thread is blocked on the queue, then it will take a mutex to interrupt the other thread in  [rb_threadptr_interrupt_common](https://github.com/ruby/ruby/blob/v2_7_0/thread.c#L491-L507).  Also, I would expect that waking up another native thread would also require another system call in the unblock function call `(th->unblock.func)(th->unblock.arg)`.  So it looks like this only gets to all happen in user-space when the queue isn't already empty.

I'm concerned the benchmark results aren't representative of real use of this library, since it seems like the performance benefit comes from queuing multiple datagrams consecutively.  If datagrams are always queued to an empty queue, then it seems like this could add extra overhead.

> I think a future improvement could be to buffer some packets and flush them out as batch

This seems like it could be a much better approach.  For instance, we could buffer datagrams during a unit of work (e.g. web request) and use [sendmmsg](https://linux.die.net/man/2/sendmmsg) to flush the buffer at the end of that unit of work or when the buffer gets full in the middle of that unit of work.