@dylanahsmith 2020-03-19
I would look at whatever stats we were looking at that led us to do this optimization.  It sounds like we found this hotspot through profiling, so I think it makes sense to compare the profile from after this to one we captured before.  Also, if we believe this was related to a performance regression that we noticed and we were hoping this would help resolve, then we could see if this helps with that metric.

Do we plan on reverting this if it has a negligible effect?