@casperisfine 2021-03-30
I took the liberty to rebase this and to add a few fixes. Notably:

### Fork handling

When forking all the thread but the main one are dead in the child process. So if we use a dispatcher thread, they need to be respawned after fork. Either automatically (e..g by checking `Process.pid` all the time, or [by decorating `Process.fork`](https://github.com/rails/rails/blob/5de2dbff16009ebf8441e259d95d4b22d068ac8c/activesupport/lib/active_support/fork_tracker.rb)) or either by requiring the users to call some method.

I went for the later and introduced a `.after_fork` method for that purpose. But I'm open to implement "fork tracking" too. It would be easier to use, but libraries monkey patching core methods are not generally well received. 

It's also important to clear the queues after fork, because any message left in them would have been inserted in the parent process, so we don't want to duplicate them.

### Make it optional

Instead of changing the existing `UPDSink` I added an extra `ThreadedUDPSink`, I feel like it's better to have it as an option at least until we experiment with it in production. We cal always make it the default later on.

### Tests

I refactored the tests a bit to be more reliable. Mostly the one about sending events from a signal handler. Since the mutex is eliminated, it's no longer a problem. However the if forked process exit too quickly after the packet is written, it won't make it to the parent. 

@methodmissing @wvanbergen @dylanahsmith what do you think?