@casperisfine 2021-03-31
> makes sense to compare the profile from after this to one we captured before.

The problem with this is that the overhead isn't really reduced, just moved in a background thread. I think a profiler would probably report a very similar time spent in the sink.